Migrate `PolicyGroup` resource to `infer`
Replaces the manual `PulumiServicePolicyGroupResource` implementation
with an infer-based resource. The Cloud API contract is unchanged:
entityType, mode, name, and organizationName still force replacement;
stacks, accounts, and policyPacks update in-place via the existing
batch endpoint; child accounts are still skipped on remove when their
parent is still in the input set.

Update sends each batch op in its own request to dodge the Cloud's
reorder/upsert-by-name behavior, recovering real state for the
checkpoint if a mid-batch op fails. Policy-pack equality compares
config, treating a nil and an empty config map as equal.

Removes the now-unused schema blocks (`PolicyGroup`,
`PolicyGroupStackReference`, `PolicyGroupPolicyPackReference`,
`PolicyGroupPolicyPackReferenceInput`) from `manual-schema.json`; the
infer framework regenerates them from the Go struct definitions. The
generated SDK `version` field is now typed as `integer` rather than
`number`, matching the API client's `int` type.

+++ sdk/dotnet/Outputs/PolicyGroupPolicyPackReference.cs
@@ -31,7 +31,7 @@ namespace Pulumi.PulumiService.Outputs
         /// <summary>
         /// The server-derived numeric version of the policy pack. This is output-only; use `versionTag` to pin a specific version.
         /// </summary>
-        public readonly double? Version;
+        public readonly int? Version;
         /// <summary>
         /// The version tag of the policy pack.
         /// </summary>
@@ -45,7 +45,7 @@ namespace Pulumi.PulumiService.Outputs
 
             string name,
 
-            double? version,
+            int? version,
 
             string? versionTag)
         {
